models/GNN-Structure2vec-GSight-master/requirements.txt
backports.weakref==1.0.post1
bleach==1.5.0
enum34==1.1.6
funcsigs==1.0.2
futures==3.1.0
html5lib==0.9999999
Markdown==2.6.10
mock==2.0.0
numpy==1.13.3
pbr==3.1.1 
protobuf==3.5.0.post1
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
sklearn==0.0
tensorflow-gpu==1.4.1
tensorflow-tensorboard==0.4.0rc3
Werkzeug==0.13
